requirements.txt
fastapi
uvicorn
firebase-admin
librosa
numpy
torchvision
pydantic
torchaudio
webrtcvad
pyworld
praat-parselmouth
google-cloud-speech